Desktop/trading bot/requirements.txt
# MT5 Trading Bot Requirements
# Install with: pip install -r requirements.txt

MetaTrader5>=5.0.45
numpy>=1.21.0
python-dotenv>=1.0.0
